feat: separate SetOnlyOpenQuestionsMode component
